implement trapping for remote remaster
bash is just so terrible to read :(

the problem with running set -e and mounts is that the mounts would not
get unbound if something between mount and unmount exist1 thus making the
next script fail as well because it can't cleanup the extractdir and so
forth.

to prevent this from happening there are now two functions.
chroot_setup which mounts and sets a trap on the usual suspects for
premature exit.
chroot_teardown which unmounts and unsets the trap again

intended result is that an exit1 between setup and teardown will always
end up calling teardown through the trap