reenable all code parts and explicitly exclude some things from removal
various ubuntu-touch related packages as well as system-image are being
explicitly installed after the removal. assuming correct dependency lineup
they should bring everything of relevance back in and result in a working
image. ubuntu-touch-session is a bit of a querstionable beast in that
regard as it also wants unity8 but appears to actually be required.
so, try without it for now and see where we get.