do not autoremove
with our latest workaroundery the image boots but isn't connectable via
adb. my theory right now is that the autoremove removes something needed
for adb. if so not autoremoving should yield a proper image now.
if not further workarounding will be necessary, and more importantly first
isolating the responsible bit -.-