only remove unity8
while yesterday a remaster with nothing but languages and lightdm removed,
and singlelogin moved to replace lightdm yielded a booting image, today
removing all seemingly irrelevant things and still replacing lightdm
yields a broken image again.
it is most confusing and annoying.
only remove unity8 and lightdm and languages to make sure it's not the
unity8 package itself that renders things broken